Issue fixed
